app/config.py
import os
from dotenv import load_dotenv

# Carrega as variáveis de ambiente do arquivo .env
basedir = os.path.abspath(os.path.dirname(__file__))
load_dotenv(os.path.join(basedir, '..', '.env')) # Aponta para o .env na raiz do projeto

class Config:
    SECRET_KEY = os.environ.get('SECRET_KEY') or 'fallback_secret_key_if_not_set'

    # Configuração SQLAlchemy - usando SQLite por padrão para desenvolvimento
    DATABASE_URL = os.environ.get('DATABASE_URL')
    if DATABASE_URL:
        SQLALCHEMY_DATABASE_URI = DATABASE_URL
    else:
        # SQLite para desenvolvimento local
        SQLALCHEMY_DATABASE_URI = f"sqlite:///{os.path.join(basedir, '..', 'carteirinha_piscina.db')}"
    
    SQLALCHEMY_TRACK_MODIFICATIONS = False

    # Configurações de Email
    MAIL_SERVER = os.environ.get('MAIL_SERVER') or 'smtp.gmail.com'
    MAIL_PORT = int(os.environ.get('MAIL_PORT') or 587)
    MAIL_USE_TLS = os.environ.get('MAIL_USE_TLS', 'true').lower() in ['true', 'on', '1']
    MAIL_USERNAME = os.environ.get('MAIL_USERNAME')
    MAIL_PASSWORD = os.environ.get('MAIL_PASSWORD')
    MAIL_DEFAULT_SENDER = os.environ.get('MAIL_DEFAULT_SENDER')

    # Configuração de upload de arquivos
    UPLOAD_FOLDER = os.path.join(basedir, 'static', 'uploads')
    MAX_CONTENT_LENGTH = 16 * 1024 * 1024  # 16MB max file size
    ALLOWED_EXTENSIONS = {'txt', 'pdf', 'png', 'jpg', 'jpeg', 'gif', 'doc', 'docx'}

    # Configuração do Scheduler
    SCHEDULER_API_ENABLED = True

    # Outras configurações da aplicação podem ser adicionadas aqui
    # Ex: MAIL_SERVER, MAIL_PORT, etc.