requirements.txt
blinker==1.9.0
click==8.2.0
colorama==0.4.6
Flask==3.1.1
Flask-WTF==1.2.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Mail==0.10.0
Flask-APScheduler==1.13.1
APScheduler==3.10.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
python-dotenv==1.1.0
Werkzeug==3.1.3
WTForms==3.2.1
email-validator==2.1.1
Pillow==10.4.0
plotly==5.24.1
pandas==2.2.3
